+++ src/configs/index.ts
@@ -10,15 +10,11 @@ export const devDashboardApiUrl = 'https://escrow-api.linkdrop.io/staging'
 
 export const nativeTokenAddress = '0x0000000000000000000000000000000000000000'
 
-export const cbwEscrowContract = '0xedfea6336c922f896c7e09ba282beb0cb4476675'
-export const mainEscrowContract = '0x139b79602b68e8198ea3d57f5e6311fd98262269'
+export const cbwEscrowContract = '0x5badb0143f69015c5c86cbd9373474a9c8ab713b'
+export const mainEscrowContract = '0xbe7b40eb3a9d85d3a76142cb637ab824f0d35ead'
 
-export const cbwEscrowContractNFT = '0xff3471dfdc6f82694e5ad4d4e7ffedf23e1e38e0'
-export const mainEscrowContractNFT = '0xe0cec4f0b66257fc6b13652c303237de0fd92ed8'
-
-export const immutableZkevmContract = '0x4366caf3963d147da4a4287061354058d871d1be'
-export const immutableZkevmContractNFT = '0x317d2501396fe75d997799bf3bdbc7cc6768b533'
-
+export const cbwEscrowContractNFT = '0x3c74782de03c0402d207fe41307fe50fe9b6b5c7'
+export const mainEscrowContractNFT = '0x5fc1316119a1b7cec52a2984c62764343dca70c9'
 
 export const MAX_MESSAGE_ENCRYPTION_KEY_LENGTH = 43
 export const MIN_MESSAGE_ENCRYPTION_KEY_LENGTH = 6
@@ -27,9 +23,7 @@ export const MAX_MESSAGE_TEXT_LENGTH = 140
 export const supportedStableCoins = {
   [ETokenAddress.usdcBase]: ESelectors.receiveWithAuthorization,
   [ETokenAddress.eurcBase]: ESelectors.receiveWithAuthorization,
-  [ETokenAddress.usdcBaseGoerli]: ESelectors.receiveWithAuthorization,
   [ETokenAddress.usdcBridgedPolygon]: ESelectors.approveWithAuthorization,
-  [ETokenAddress.usdcSepolia]: ESelectors.receiveWithAuthorization,
   [ETokenAddress.usdcPolygon]: ESelectors.receiveWithAuthorization,
   [ETokenAddress.usdcArbitrum]: ESelectors.receiveWithAuthorization,
   [ETokenAddress.usdcOptimism]: ESelectors.receiveWithAuthorization,