+++ src/domains/index.ts
@@ -3,9 +3,7 @@ import polygonUSDCBridged from "./polygon-bridged-usdc"
 import optimismUSDC from "./optimism-usdc"
 import avalancheUSDC from "./avalanche-usdc"
 import arbitrumUSDC from "./arbitrum-usdc"
-import baseGoerliUSDC from "./base-goerli-usdc"
 import baseUSDC from "./base-usdc"
-import sepoliaUSDC from "./sepolia-usdc"
 import baseEURC from "./base-eurc"
 import baseCBBTC from "./base-cbbtc"
 
@@ -17,7 +15,5 @@ export {
   optimismUSDC,
   avalancheUSDC,
   arbitrumUSDC,
-  baseGoerliUSDC,
-  baseUSDC,
-  sepoliaUSDC
+  baseUSDC
 }