+++ src/modules/claim-link/index.ts
@@ -238,8 +238,7 @@ class ClaimLink implements IClaimLinkSDK {
       if (!defineIfEscrowAddressIsCorrect(
         this.escrowAddress as string,
         this.tokenType,
-        this.deployment,
-        this.chainId
+        this.deployment
       )) {
         throw new Error(errors.escrow_is_not_correct())
       }
@@ -319,7 +318,8 @@ class ClaimLink implements IClaimLinkSDK {
         this.expiration,
         this.feeToken,
         this.feeAmount,
-        this.feeAuthorization
+        this.feeAuthorization,
+        this.encryptedSenderMessage ? `0x${this.encryptedSenderMessage}` : '0x'
       ])
     } else if (this.tokenType === 'NATIVE') {
       const iface = new ethers.Interface(LinkdropEscrowToken.abi)
@@ -328,7 +328,8 @@ class ClaimLink implements IClaimLinkSDK {
         this.totalAmount,
         this.expiration,
         this.feeAmount,
-        this.feeAuthorization
+        this.feeAuthorization,
+        this.encryptedSenderMessage ? `0x${this.encryptedSenderMessage}` : '0x'
       ])
     } else if (this.tokenType === 'ERC721') {
       const iface = new ethers.Interface(LinkdropEscrowNFT.abi)
@@ -338,7 +339,8 @@ class ClaimLink implements IClaimLinkSDK {
         this.tokenId,
         this.expiration,
         this.feeAmount,
-        this.feeAuthorization
+        this.feeAuthorization,
+        this.encryptedSenderMessage ? `0x${this.encryptedSenderMessage}` : '0x'
       ])
     } else {
       const iface = new ethers.Interface(LinkdropEscrowNFT.abi)
@@ -349,7 +351,8 @@ class ClaimLink implements IClaimLinkSDK {
         this.amount,
         this.expiration,
         this.feeAmount,
-        this.feeAuthorization
+        this.feeAuthorization,
+        this.encryptedSenderMessage ? `0x${this.encryptedSenderMessage}` : '0x'
       ])
     }